requirements.txt
asgiref==3.2.7
autopep8==1.5.2
Django==3.0.5
django-friendship==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
entrypoints==0.3
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
